requirements.txt
# Extracted from environment.yml (pip section)
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beartype==0.14.1
better-abc==0.0.3
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
cmake==3.25.0
datasets==2.17.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
fancy-einsum==0.0.3
filelock==3.9.0
frozenlist==1.4.1
fsspec==2023.10.0
gitdb==4.0.11
gitpython==3.1.42
huggingface-hub==0.21.1
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.4
jaxtyping==0.2.25
jinja2==3.1.2
lit==15.0.7
markdown-it-py==3.0.0
markupsafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.1
numpy==1.24.1
omegaconf==2.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.0
pyarrow-hotfix==0.6
pygments==2.17.2
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
regex==2023.12.25
requests==2.28.1
rich==13.7.0
safetensors==0.4.2
sentry-sdk==1.40.6
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
submitit==1.5.1
sympy==1.12
tokenizers==0.15.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.2
transformer-lens==1.14.0
transformers==4.38.1
triton==2.0.0
typeguard==2.13.3
typing-extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.16.3
xxhash==3.4.1
yarl==1.9.4
